Chore: 사용 가능 기프티콘 정렬 조건 추가

Closes #125

See merge request s12-blockchain-nft-sub1/S12P21E101!184

+++ backend/src/main/java/com/e101/nift/secondhand/model/dto/response/ArticleDetailDto.java
@@ -32,5 +32,5 @@ public class ArticleDetailDto {
 
     private String userNickName;
     private String profileImage;
-    private Integer sellerTxs; // 판매자가 참여한 거래수
+    private Long sellerTxs; // 판매자가 참여한 거래수
 }
